+++ command/build_test.go
@@ -67,7 +67,7 @@ func TestBuild(t *testing.T) {
 		},
 
 		{
-			name: "var-args: json - inexistent var file errs",
+			name: "var-args: json - nonexistent var file errs",
 			args: []string{
 				"-var-file=" + filepath.Join(testFixture("var-arg"), "potato.json"),
 				filepath.Join(testFixture("var-arg"), "fruit_builder.json"),
@@ -77,7 +77,7 @@ func TestBuild(t *testing.T) {
 		},
 
 		{
-			name: "var-args: hcl - inexistent json var file errs",
+			name: "var-args: hcl - nonexistent json var file errs",
 			args: []string{
 				"-var-file=" + filepath.Join(testFixture("var-arg"), "potato.json"),
 				testFixture("var-arg"),
@@ -87,7 +87,7 @@ func TestBuild(t *testing.T) {
 		},
 
 		{
-			name: "var-args: hcl - inexistent hcl var file errs",
+			name: "var-args: hcl - nonexistent hcl var file errs",
 			args: []string{
 				"-var-file=" + filepath.Join(testFixture("var-arg"), "potato.hcl"),
 				testFixture("var-arg"),
@@ -404,6 +404,25 @@ func TestBuild(t *testing.T) {
 					"example2.1.txt",
 				},
 			},
+		},
+		{
+			name: "hcl - using build variables in post-processor",
+			args: []string{
+				testFixture("hcl", "build-var-in-pp.pkr.hcl"),
+			},
+			fileCheck: fileCheck{
+				expectedContent: map[string]string{
+					"example.2.txt": two,
+				},
+			},
+		},
+
+		{
+			name: "hcl - test crash #11381",
+			args: []string{
+				testFixture("hcl", "nil-component-crash.pkr.hcl"),
+			},
+			expectedCode: 1,
 		},
 	}