+++ hcl2template/types.build.provisioners.go
@@ -122,6 +122,7 @@ func (p *Parser) decodeProvisioner(block *hcl.Block, cfg *PackerConfig) (*Provis
 				Summary:  "Failed to parse pause_before duration",
 				Severity: hcl.DiagError,
 				Detail:   err.Error(),
+				Subject:  &block.DefRange,
 			})
 		}
 		provisioner.PauseBefore = pauseBefore
@@ -131,8 +132,10 @@ func (p *Parser) decodeProvisioner(block *hcl.Block, cfg *PackerConfig) (*Provis
 		timeout, err := time.ParseDuration(b.Timeout)
 		if err != nil {
 			return nil, append(diags, &hcl.Diagnostic{
-				Summary: "Failed to parse timeout duration",
-				Detail:  err.Error(),
+				Summary:  "Failed to parse timeout duration",
+				Severity: hcl.DiagError,
+				Detail:   err.Error(),
+				Subject:  &block.DefRange,
 			})
 		}
 		provisioner.Timeout = timeout