+++ hcl2template/types.source.go
@@ -110,6 +110,9 @@ func (cfg *PackerConfig) startBuilder(source SourceUseBlock, ectx *hcl.EvalConte
 	}
 
 	body := source.Body
+	// Add known values to source accessor in eval context.
+	ectx.Variables[sourcesAccessor] = cty.ObjectVal(source.ctyValues())
+
 	decoded, moreDiags := decodeHCL2Spec(body, ectx, builder)
 	diags = append(diags, moreDiags...)
 	if moreDiags.HasErrors() {